+++ processor/src/test/resources/bad_input/com/example/cascade_dispose/BadType1Component.java
@@ -1,0 +1,12 @@
+package com.example.cascade_dispose;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.CascadeDispose;
+
+@ArezComponent
+public abstract class BadType1Component
+{
+  // Should be declared type (a.k.a Class type)
+  @CascadeDispose
+  int _myField;
+}