+++ processor/src/test/resources/bad_input/com/example/cascade_dispose/BadType3Component.java
@@ -1,0 +1,13 @@
+package com.example.cascade_dispose;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.CascadeDispose;
+import java.util.List;
+
+@ArezComponent
+public abstract class BadType3Component
+{
+  // Class does not implement Disposable
+  @CascadeDispose
+  List _myField;
+}