+++ processor/src/test/resources/bad_input/com/example/package_access/other/BaseCascadeDisposeModel.java
@@ -1,0 +1,10 @@
+package com.example.package_access.other;
+
+import arez.Disposable;
+import arez.annotations.CascadeDispose;
+
+public abstract class BaseCascadeDisposeModel
+{
+  @CascadeDispose
+  Disposable _myDisposableField;
+}