+++ processor/src/test/resources/input/com/example/cascade_dispose/ComponentCascadeDisposeModel.java
@@ -1,0 +1,16 @@
+package com.example.cascade_dispose;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.CascadeDispose;
+
+@ArezComponent
+public abstract class ComponentCascadeDisposeModel
+{
+  @CascadeDispose
+  protected MyComponent _myElement;
+
+  @ArezComponent( allowEmpty = true )
+  static abstract class MyComponent
+  {
+  }
+}